in-memory queue syncs with strapi queue
